// src/types/survey.ts
export interface SurveyData {
  surveyType: 'adult' | 'child' | '';
  userName: string;
  birthDate: string;
  age: string;
  diagnosis: string;
  modality: string[];
  applicationDate: string;
  professional: string;
  caregiverContact: string;
  updateReasons: string[];
  updateReasonDescription: string;
  livesWith: string;
  spendsMostTimeWith: string;
  organizesRoutine: string;
  accompaniesDisplacements: string;
  routineMorning: string;
  routineMainPeriod: string;
  routineAfternoonNight: string;
  functionalityOptions: string[];
  mainDifficulties: string;
  primaryCaregiver: string;
  otherFamilySupport: string;
  treatmentFacilitators: string;
  treatmentBarriers: string;
  healthNetworkOptions: string[];
  healthNetworkObservations: string;
  occupationalNetworkOptions: string[];
  occupationalNetworkObservations: string;
  logisticalSupportOptions: string[];
  routineSynthesis: string;
  ecomapSynthesis: string;
  familyComposition: string;
// ...
  childLivesWith: string;
// ...
  childPrimaryCaregiver: string;
// ...
  childFather: string;
// ...
}